+++ scripts/point_party/eval_celltype_j0251.py
@@ -261,7 +261,7 @@ if __name__ == '__main__':
         target_names = [int2str_label[kk] for kk in range(nclasses)]
         csv_p = ccd.csv_p
         df = pandas.io.parsers.read_csv(csv_p, header=None, names=['ID', 'type']).values
-        ssv_ids = df[:, 0].astype(np.uint)
+        ssv_ids = df[:, 0].astype(np.uint64)
         if len(np.unique(ssv_ids)) != len(ssv_ids):
             raise ValueError('Multi-usage of IDs!')
         str_labels = df[:, 1]
